Fix the names for a few metrics
The big change in this commit is to fix a big mistake in wMetrics: the pressure
was mapped to some other metrics and the 'w' is actually 'y'. :-)

The other change was to update the metrics name for staleness of ms_v_env_temp
to ms_v_env_temp_indicator to avoid the collision with the real value.